pyproject.toml
[project]
name = "user-management"
version = "0.1.0"
description = "JWT-authenticated user management API with FastAPI and React"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.136.0",
    "pydantic[email]==2.13.2",
    "sqlalchemy==2.0.49",
    "alembic==1.18.4",
    "uvicorn[standard]==0.44.0",
    "email-validator==2.3.0",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI uses function calls in default args (Depends)

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
